requirements.txt
streamlit
requests
openai
python-dotenv
PyMuPDF
python-docx
elevenlabs
bs4
qdrant-client
fastembed
firecrawl
